Summoner model repository backed by RIOT API

+++ app/src/main/java/com/tsarsprocket/reportmid/presentation/PlayerPresentation.kt
@@ -1,17 +1,15 @@
 package com.tsarsprocket.reportmid.presentation
 
-import android.graphics.Bitmap
 import android.graphics.drawable.Drawable
 import androidx.lifecycle.MutableLiveData
-import com.tsarsprocket.reportmid.model.ChampionModel
-import com.tsarsprocket.reportmid.model.SummonerModel
+import com.tsarsprocket.reportmid.summoner.model.SummonerModel
 
 class PlayerPresentation() {
     val summoner = MutableLiveData<SummonerModel>()
     val championIconLive = MutableLiveData<Drawable>()
     val summonerChampionSkillLive = MutableLiveData<Int>()
     val summonerNameLive = MutableLiveData<String>()
-    val summonerLevelLive = MutableLiveData<Int>()
+    val summonerLevelLive = MutableLiveData<Long>()
     val soloqueueRankLive = MutableLiveData<String>()
     val soloqueueWinrateLive = MutableLiveData<Float>()
     val summonerSpellDLive = MutableLiveData<Drawable>()
